move "reactions support for messages" to Watchman.Discord.
