+++ Watchman.Discord/Program.cs
@@ -3,6 +3,7 @@ using Newtonsoft.Json;
 using Serilog;
 using System.IO;
 using System.Threading.Tasks;
+using Watchman.Discord.Areas.Reactions.Services;
 
 namespace Watchman.Discord
 {
@@ -21,6 +22,8 @@ namespace Watchman.Discord
             
             var watchman = new WatchmanBot(configuration);
             var workflowBuilder = watchman.GetWorkflowBuilder();
+            workflowBuilder.ReactionAdded += (userMessage, socketMessageChannel, socketReaction) => ReactionsService.OnReactionAddedEvent(socketReaction);
+            workflowBuilder.ReactionRemoved += (userMessage, socketMessageChannel, socketReaction) => ReactionsService.OnReactionRemovedEvent(socketReaction);
             var lastException = DateTime.MinValue;
             var numberOfExceptionsInLastTime = 0;
